OOP_Project_Faculty_institute_configuration
In this code, we have two classes: Faculty and Institute. The Faculty class represents an individual faculty member with a name and department, while the Institute class represents an academic institution with a name, location, and a list of faculties.

The add_faculty method allows adding a faculty member to the institute's list of faculties, while the remove_faculty method removes a faculty member from the list. The print_faculties method is used to print out the name and department of all faculty members in the institute.